# services/llm-api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OpenAI-compatible LLM API platform with enterprise authentication, conversation management, and streaming support.",
        "title": "Jan Server LLM API",
        "contact": {
            "name": "Jan Server Team",
            "url": "https://github.com/janhq/jan-server"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of available models that can be used for chat completions or other tasks. Returns either simple model list or detailed list with provider metadata based on X-PROVIDER-DATA header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat Completions API"
                ],
                "summary": "List available models",
                "parameters": [
                    {
                        "enum": [
                            "true",
                            "false"
                        ],
                        "type": "string",
                        "description": "Set to 'true' to include provider metadata in response",
                        "name": "X-PROVIDER-DATA",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of models with provider metadata (when X-PROVIDER-DATA=true)",
                        "schema": {
                            "$ref": "#/definitions/modelresponses.ModelWithProviderResponseList"
                        }
                    },
                    "404": {
                        "description": "Models or providers not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve models",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/models/catalogs/{model_public_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a model catalog entry by its public ID (supports IDs with slashes like openrouter/nova-lite-v1)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model API"
                ],
                "summary": "Get a model catalog entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model Catalog Public ID (can contain slashes)",
                        "name": "model_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model catalog details",
                        "schema": {
                            "$ref": "#/definitions/modelresponses.ModelCatalogResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Model catalog not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Architecture": {
            "type": "object",
            "properties": {
                "input_modalities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instruct_type": {
                    "description": "nullable",
                    "type": "string"
                },
                "modality": {
                    "description": "\"text+image-\u003etext\"",
                    "type": "string"
                },
                "output_modalities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokenizer": {
                    "description": "\"GPT\" / \"SentencePiece\" / etc.",
                    "type": "string"
                }
            }
        },
        "model.ModelCatalogStatus": {
            "type": "string",
            "enum": [
                "init",
                "filled",
                "updated"
            ],
            "x-enum-comments": {
                "ModelCatalogStatusFilled": "may update from Provider like OpenRouter",
                "ModelCatalogStatusInit": "default status when creating entry",
                "ModelCatalogStatusUpdated": "manually updated by admin (cannot be auto-updated anymore"
            },
            "x-enum-descriptions": [
                "default status when creating entry",
                "may update from Provider like OpenRouter",
                "manually updated by admin (cannot be auto-updated anymore"
            ],
            "x-enum-varnames": [
                "ModelCatalogStatusInit",
                "ModelCatalogStatusFilled",
                "ModelCatalogStatusUpdated"
            ]
        },
...
        "model.SupportedParameters": {
            "type": "object",
            "properties": {
                "default": {
                    "description": "temperature/top_p/frequency_penalty, null allowed",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                },
                "names": {
                    "description": "e.g., [\"include_reasoning\",\"max_tokens\",...]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "modelresponses.ModelCatalogResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "architecture": {
                    "$ref": "#/definitions/model.Architecture"
                },
                "created_at": {
                    "type": "integer"
                },
                "extras": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "type": "string"
                },
                "is_moderated": {
                    "type": "boolean"
                },
                "last_synced_at": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.ModelCatalogStatus"
                },
                "supported_parameters": {
                    "$ref": "#/definitions/model.SupportedParameters"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
        "modelresponses.ModelResponseWithProvider": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "string"
                },
                "provider_name": {
                    "type": "string"
                },
                "provider_vendor": {
                    "type": "string"
                }
            }
        },
        "modelresponses.ModelWithProviderResponseList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelresponses.ModelResponseWithProvider"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "UUID from PlatformError",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}